moved the text data model into its own module
