+++ tests/samples/misc.py
@@ -1,0 +1,1 @@
+FAKE_LINK = "https://en.wikipedia.org/wiki/Blue_whale"
